Locomotion/ai_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
qdrant-client==1.9.1
sentence-transformers==3.0.1
google-generativeai==0.8.3
pydantic==2.7.4
python-dotenv==1.0.1
langchain
langchain-google-genai
langchain-community
langgraph
langchain-core
